Move relevant classes to render package
