+++ src/main/java/bdv/fx/viewer/render/MultiResolutionRendererFX.java
@@ -27,7 +27,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package bdv.fx.viewer;
+package bdv.fx.viewer.render;
 
 import java.util.concurrent.ExecutorService;
