+++ src/main/java/bdv/fx/viewer/render/TransformAwareBufferedImageOverlayRendererGeneric.java
@@ -1,13 +1,9 @@
 /*
  * #%L
- * ImgLib2: a general-purpose, multidimensional image processing library.
+ * BigDataViewer core classes with minimal dependencies
  * %%
- * Copyright (C) 2009 - 2016 Tobias Pietzsch, Stephan Preibisch, Stephan Saalfeld,
- * John Bogovic, Albert Cardona, Barry DeZonia, Christian Dietz, Jan Funke,
- * Aivar Grislis, Jonathan Hale, Grant Harris, Stefan Helfrich, Mark Hiner,
- * Martin Horn, Steffen Jaensch, Lee Kamentsky, Larry Lindsey, Melissa Linkert,
- * Mark Longair, Brian Northan, Nick Perry, Curtis Rueden, Johannes Schindelin,
- * Jean-Yves Tinevez and Michael Zinsmaier.
+ * Copyright (C) 2012 - 2016 Tobias Pietzsch, Stephan Saalfeld, Stephan Preibisch,
+ * Jean-Yves Tinevez, HongKee Moon, Johannes Schindelin, Curtis Rueden, John Bogovic
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -31,28 +27,10 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package bdv.fx.viewer;
+package bdv.fx.viewer.render;
 
-/**
- * Draw something to a canvas and receive notifications about changes of the canvas size.
- *
- * @author Tobias Pietzsch
- */
-public interface OverlayRendererGeneric<G>
+public interface TransformAwareBufferedImageOverlayRendererGeneric<G, T>
+		extends OverlayRendererGeneric<G>, RenderTargetGeneric<T>, TransformAwareRenderTargetGeneric<T>
 {
-	/**
-	 * Render overlays.
-	 */
-	public void drawOverlays(final G g);
 
-	/**
-	 * This is called, when the screen size of the canvas (the component displaying the image and generating mouse
-	 * events) changes. This can be used to determine scale of overlay or screen coordinates relative to the border.
-	 *
-	 * @param width
-	 * 		the new canvas width.
-	 * @param height
-	 * 		the new canvas height.
-	 */
-	public void setCanvasSize(final int width, final int height);
 }
