+++ src/main/java/bdv/fx/viewer/render/TransformAwareRenderTargetGeneric.java
@@ -27,10 +27,24 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package bdv.fx.viewer;
+package bdv.fx.viewer.render;
 
-public interface TransformAwareBufferedImageOverlayRendererGeneric<G, T>
-		extends OverlayRendererGeneric<G>, RenderTargetGeneric<T>, TransformAwareRenderTargetGeneric<T>
+import net.imglib2.realtransform.AffineTransform3D;
+import net.imglib2.ui.TransformListener;
+
+public interface TransformAwareRenderTargetGeneric<T> extends RenderTargetGeneric<T>
 {
+	/**
+	 * Set the data store that is to be drawn on the canvas, and the transform with which this image was created.
+	 *
+	 * @param img
+	 * 		image to draw (may be null).
+	 */
+	public T setBufferedImageAndTransform(final T img, final AffineTransform3D transform);
 
+	public void addTransformListener(final TransformListener<AffineTransform3D> listener);
+
+	public void addTransformListener(final TransformListener<AffineTransform3D> listener, final int index);
+
+	public void removeTransformListener(final TransformListener<AffineTransform3D> listener);
 }
