+++ src/main/java/org/janelia/saalfeldlab/paintera/control/paint/BrushOverlay.java
@@ -3,7 +3,7 @@ package org.janelia.saalfeldlab.paintera.control.paint;
 import java.lang.invoke.MethodHandles;
 import java.util.stream.IntStream;
 
-import bdv.fx.viewer.OverlayRendererGeneric;
+import bdv.fx.viewer.render.OverlayRendererGeneric;
 import bdv.fx.viewer.ViewerPanelFX;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.SimpleDoubleProperty;
