+++ src/main/java/org/janelia/saalfeldlab/paintera/control/paint/FillOverlay.java
@@ -2,7 +2,7 @@ package org.janelia.saalfeldlab.paintera.control.paint;
 
 import java.lang.invoke.MethodHandles;
 
-import bdv.fx.viewer.OverlayRendererGeneric;
+import bdv.fx.viewer.render.OverlayRendererGeneric;
 import bdv.fx.viewer.ViewerPanelFX;
 import javafx.scene.Cursor;
 import javafx.scene.canvas.GraphicsContext;
